Refactor ASAR patcher and validate required patches
