+++ scripts/patches/computer-use.js
@@ -1,0 +1,237 @@
+"use strict";
+
+const fs = require("node:fs");
+const path = require("node:path");
+
+const COMPUTER_USE_UI_ENV_VAR = "CODEX_LINUX_ENABLE_COMPUTER_USE_UI";
+const COMPUTER_USE_UI_SETTINGS_KEY = "codex-linux-computer-use-ui-enabled";
+
+// Computer Use has two postures: the bundled plugin gate is default-on Linux
+// platform glue; the visible UI gates remain opt-in because they bypass rollout
+// checks in upstream webview code.
+function isComputerUseUiEnabled(env = process.env) {
+  if (env[COMPUTER_USE_UI_ENV_VAR] === "1") {
+    return true;
+  }
+  return readComputerUseUiSettingsFlag(env);
+}
+
+function readComputerUseUiSettingsFlag(env) {
+  const settingsPath = computerUseUiSettingsPath(env);
+  if (settingsPath == null) {
+    return false;
+  }
+  try {
+    if (!fs.existsSync(settingsPath)) {
+      return false;
+    }
+    const raw = fs.readFileSync(settingsPath, "utf8");
+    const parsed = JSON.parse(raw);
+    if (parsed == null || typeof parsed !== "object" || Array.isArray(parsed)) {
+      return false;
+    }
+    return parsed[COMPUTER_USE_UI_SETTINGS_KEY] === true;
+  } catch {
+    return false;
+  }
+}
+
+function computerUseUiSettingsPath(env) {
+  const xdgConfig = env.XDG_CONFIG_HOME;
+  const home = env.HOME;
+  const configHome = (xdgConfig && xdgConfig.length > 0)
+    ? xdgConfig
+    : home
+      ? path.join(home, ".config")
+      : null;
+  return configHome == null ? null : path.join(configHome, "codex-desktop", "settings.json");
+}
+
+// Lookback/lookahead windows used when searching for the nearest minified
+// identifier or surrounding context around a regex anchor in the bundle.
+// Sized empirically to the typical distance between a feature's anchor and
+// the helper aliases it depends on.
+const TRAY_GUARD_LOOKAHEAD = 1200;
+const CLOSE_GATE_PREFIX_LOOKBACK = 8000;
+const HANDLER_PREFIX_LOOKBACK = 12000;
+const DIRECT_HANDLER_PROXIMITY = 1200;
+
+const linuxSettingsKeys = {
+  promptWindow: "codex-linux-prompt-window-enabled",
+  systemTray: "codex-linux-system-tray-enabled",
+  warmStart: "codex-linux-warm-start-enabled",
+};
+
+function parseDestructuredParamAliases(paramsText) {
+  const aliases = Object.create(null);
+  for (const rawPart of paramsText.split(",")) {
+    const part = rawPart.trim();
+    const match = part.match(/^([A-Za-z_$][\w$]*)(?::([A-Za-z_$][\w$]*))?$/);
+    if (match != null) {
+      aliases[match[1]] = match[2] ?? match[1];
+    }
+  }
+  return aliases;
+}
+
+function buildComputerUseGate({ nameExpr, availabilityProp, featuresVar, platformVar, migrateVar }) {
+  return `{installWhenMissing:!0,name:${nameExpr},${availabilityProp}:({features:${featuresVar},platform:${platformVar}})=>(${platformVar}===\`darwin\`||${platformVar}===\`linux\`)&&${featuresVar}.computerUse,migrate:${migrateVar}}`;
+}
+
+function hasComputerUseLiteral(source) {
+  return /(?:`computer-use`|"computer-use"|'computer-use')/.test(source);
+}
+
+function isComputerUseNameExpr(nameExpr, computerUseNameVar) {
+  return /^(?:`computer-use`|"computer-use"|'computer-use')$/.test(nameExpr) || nameExpr === computerUseNameVar;
+}
+
+function applyLinuxComputerUsePluginGatePatch(currentSource) {
+  if (!hasComputerUseLiteral(currentSource)) {
+    console.warn(
+      "WARN: Could not find Computer Use plugin gate literal — skipping Linux Computer Use plugin gate patch",
+    );
+    return currentSource;
+  }
+
+  const computerUseNameVar = currentSource.match(/([A-Za-z_$][\w$]*)=(?:`computer-use`|"computer-use"|'computer-use')/)?.[1] ?? null;
+  const gateRegex =
+    /\{(installWhenMissing:!0,)?name:([A-Za-z_$][\w$]*|`computer-use`|"computer-use"|'computer-use'),(isEnabled|isAvailable):\(\{([^}]*)\}\)=>([^{}]*?\.computerUse),migrate:([A-Za-z_$][\w$]*)\}/g;
+  let sawEnabledGate = false;
+  let sawUnpatchableGate = false;
+  let match;
+  while ((match = gateRegex.exec(currentSource)) != null) {
+    const [gateSource, installWhenMissing, nameExpr, availabilityProp, paramsText, expression, migrateVar] = match;
+    if (!isComputerUseNameExpr(nameExpr, computerUseNameVar)) {
+      continue;
+    }
+
+    const aliases = parseDestructuredParamAliases(paramsText);
+    const featuresVar = aliases.features;
+    const platformVar = aliases.platform;
+    if (featuresVar == null || platformVar == null) {
+      continue;
+    }
+
+    const darwinOnlyExpression = `${platformVar}===\`darwin\`&&${featuresVar}.computerUse`;
+    const linuxExpression = `(${platformVar}===\`darwin\`||${platformVar}===\`linux\`)&&${featuresVar}.computerUse`;
+    if (installWhenMissing != null && expression === linuxExpression) {
+      sawEnabledGate = true;
+      continue;
+    }
+    if (expression === darwinOnlyExpression || expression === linuxExpression) {
+      const replacement = buildComputerUseGate({ nameExpr, availabilityProp, featuresVar, platformVar, migrateVar });
+      return `${currentSource.slice(0, match.index)}${replacement}${currentSource.slice(match.index + gateSource.length)}`;
+    }
+    sawUnpatchableGate = true;
+  }
+
+  if (sawEnabledGate && !sawUnpatchableGate) {
+    return currentSource;
+  }
+
+  if (hasComputerUseLiteral(currentSource) && currentSource.includes("computerUse")) {
+    throw new Error("Required Linux Computer Use plugin gate patch failed: could not enable bundled Computer Use on Linux");
+  }
+
+  return currentSource;
+}
+
+function applyLinuxComputerUseFeaturePatch(currentSource) {
+  const patchedFeaturePattern =
+    /function [A-Za-z_$][\w$]*\([A-Za-z_$][\w$]*,\{env:[A-Za-z_$][\w$]*=process\.env,platform:[A-Za-z_$][\w$]*=process\.platform\}=\{\}\)\{return [A-Za-z_$][\w$]*===`linux`\?\{\.\.\.[A-Za-z_$][\w$]*,computerUse:!0,computerUseNodeRepl:!0\}:/;
+  const windowsOnlyFeaturePattern =
+    /function ([A-Za-z_$][\w$]*)\(([A-Za-z_$][\w$]*),\{env:([A-Za-z_$][\w$]*)=process\.env,platform:([A-Za-z_$][\w$]*)=process\.platform\}=\{\}\)\{return \4!==`win32`\|\|\3\.CODEX_ELECTRON_ENABLE_WINDOWS_COMPUTER_USE!==`1`\?\2:\{\.\.\.\2,computerUse:!0,computerUseNodeRepl:!0\}\}/;
+
+  if (patchedFeaturePattern.test(currentSource)) {
+    return currentSource;
+  }
+
+  if (windowsOnlyFeaturePattern.test(currentSource)) {
+    return currentSource.replace(
+      windowsOnlyFeaturePattern,
+      (_, fnName, featuresVar, envVar, platformVar) =>
+        `function ${fnName}(${featuresVar},{env:${envVar}=process.env,platform:${platformVar}=process.platform}={}){return ${platformVar}===\`linux\`?{...${featuresVar},computerUse:!0,computerUseNodeRepl:!0}:${platformVar}!==\`win32\`||${envVar}.CODEX_ELECTRON_ENABLE_WINDOWS_COMPUTER_USE!==\`1\`?${featuresVar}:{...${featuresVar},computerUse:!0,computerUseNodeRepl:!0}}`,
+    );
+  }
+
+  if (currentSource.includes("CODEX_ELECTRON_ENABLE_WINDOWS_COMPUTER_USE")) {
+    console.warn(
+      "WARN: Could not find Computer Use desktop feature gate — skipping Linux Computer Use feature patch",
+    );
+  }
+
+  return currentSource;
+}
+
+function applyLinuxComputerUseRendererAvailabilityPatch(currentSource) {
+  let patchedSource = currentSource;
+
+  const platformPredicateNeedle = "function hae(e){return e===`macOS`||e===`windows`}";
+  const platformPredicatePatch =
+    "function hae(e){return e===`macOS`||e===`windows`||e===`linux`}";
+  if (patchedSource.includes(platformPredicatePatch)) {
+    // Already patched.
+  } else if (patchedSource.includes(platformPredicateNeedle)) {
+    patchedSource = patchedSource.replace(platformPredicateNeedle, platformPredicatePatch);
+  }
+
+  const availabilityNeedle =
+    "let m=a&&i&&s===`electron`&&u&&(c||p),h=m&&!c&&f.enabled&&!f.isLoading,g=m&&f.isLoading,_=m&&(c||f.isLoading),v;";
+  const availabilityHostLocalLinuxPatch =
+    "let m=a&&i&&s===`electron`&&(l===`linux`||u&&(c||p)),h=m&&!c&&(l===`linux`||f.enabled)&&!f.isLoading,g=m&&l!==`linux`&&f.isLoading,_=m&&(c||l!==`linux`&&f.isLoading),v;";
+  const availabilityPatch =
+    "let m=a&&(i||l===`linux`)&&s===`electron`&&(l===`linux`||u&&(c||p)),h=m&&!c&&(l===`linux`||f.enabled)&&!f.isLoading,g=m&&l!==`linux`&&f.isLoading,_=m&&(c||l!==`linux`&&f.isLoading),v;";
+  if (patchedSource.includes(availabilityPatch)) {
+    return patchedSource;
+  }
+
+  if (patchedSource.includes(availabilityHostLocalLinuxPatch)) {
+    return patchedSource.replace(availabilityHostLocalLinuxPatch, availabilityPatch);
+  }
+
+  if (patchedSource.includes(availabilityNeedle)) {
+    return patchedSource.replace(availabilityNeedle, availabilityPatch);
+  }
+
+  if (currentSource.includes("featureName:`computer_use`") && currentSource.includes("isComputerUseAvailable")) {
+    console.warn(
+      "WARN: Could not find Computer Use renderer availability gate — skipping Linux Computer Use UI availability patch",
+    );
+  }
+
+  return patchedSource;
+}
+
+function applyLinuxComputerUseInstallFlowPatch(currentSource) {
+  const availabilityNeedle =
+    "ne=f({featureName:`computer_use`,hostId:t}),re=!ne.isLoading&&ne.enabled,";
+  const availabilityPatch =
+    "ne=f({featureName:`computer_use`,hostId:t}),re=!ne.isLoading&&ne.enabled||navigator.userAgent.includes(`Linux`),";
+
+  if (currentSource.includes(availabilityPatch)) {
+    return currentSource;
+  }
+
+  if (currentSource.includes(availabilityNeedle)) {
+    return currentSource.replace(availabilityNeedle, availabilityPatch);
+  }
+
+  if (currentSource.includes("featureName:`computer_use`")) {
+    console.warn(
+      "WARN: Could not find Computer Use install flow gate — skipping Linux Computer Use install flow patch",
+    );
+  }
+
+  return currentSource;
+}
+
+module.exports = {
+  COMPUTER_USE_UI_ENV_VAR,
+  COMPUTER_USE_UI_SETTINGS_KEY,
+  applyLinuxComputerUseFeaturePatch,
+  applyLinuxComputerUseInstallFlowPatch,
+  applyLinuxComputerUsePluginGatePatch,
+  applyLinuxComputerUseRendererAvailabilityPatch,
+  isComputerUseUiEnabled,
+};
