+++ scripts/patches/keybinds-settings.js
@@ -1,0 +1,344 @@
+"use strict";
+
+const fs = require("node:fs");
+const path = require("node:path");
+const {
+  findImportedAsset,
+  findRequiredWebviewAsset,
+  linuxSettingsKeys,
+} = require("./shared.js");
+
+// Keybind settings are transactional: either all dependent webview assets are
+// updated together, or the patch logs a warning and leaves the app usable.
+const keybindsSettingsAsset = "keybinds-settings-linux.js";
+const linuxKeybindOverridesKey = "codex-linux-keybind-overrides";
+
+function buildKeybindsSettingsSource({
+  chunkAsset,
+  reactAsset,
+  jsxRuntimeAsset,
+  vscodeApiAsset,
+  hotkeySettingsAsset,
+  toggleAsset,
+  settingsRowAsset,
+  settingsLayoutAsset,
+}) {
+  const keybindGroups = [
+    {
+      title: "Core",
+      actions: [
+        { id: "newThread", label: "New chat", description: "Start a new chat." },
+        { id: "quickChat", label: "Quick chat", description: "Open a quick chat window." },
+        { id: "newThreadAlt", label: "New chat alternate", description: "Alternate shortcut for a new chat." },
+        { id: "openFolder", label: "Open folder", description: "Open a workspace folder." },
+        { id: "settings", label: "Settings", description: "Open settings." },
+        { id: "openCommandMenu", label: "Command menu", description: "Open the command menu." },
+        { id: "openCommandMenuAlt", label: "Command menu alternate", description: "Alternate shortcut for the command menu." },
+        { id: "searchChats", label: "Search chats", description: "Search existing chats." },
+        { id: "searchFiles", label: "Search files", description: "Search files in the current workspace." },
+        { id: "newWindow", label: "New window", description: "Open a new app window." },
+      ],
+    },
+    {
+      title: "Thread",
+      actions: [
+        { id: "findInThread", label: "Find in thread", description: "Search inside the current thread." },
+        { id: "copyConversationPath", label: "Copy conversation path", description: "Copy the current conversation path." },
+        { id: "toggleThreadPin", label: "Toggle thread pin", description: "Pin or unpin the current thread." },
+        { id: "renameThread", label: "Rename thread", description: "Rename the current thread." },
+        { id: "archiveThread", label: "Archive thread", description: "Archive the current thread." },
+        { id: "copyWorkingDirectory", label: "Copy working directory", description: "Copy the current working directory." },
+        { id: "copySessionId", label: "Copy session ID", description: "Copy the current session ID." },
+        { id: "copyDeeplink", label: "Copy deeplink", description: "Copy a deeplink for the current thread." },
+        { id: "previousThread", label: "Previous thread", description: "Move to the previous thread." },
+        { id: "nextThread", label: "Next thread", description: "Move to the next thread." },
+        { id: "thread1", label: "Thread 1", description: "Jump to thread slot 1." },
+        { id: "thread2", label: "Thread 2", description: "Jump to thread slot 2." },
+        { id: "thread3", label: "Thread 3", description: "Jump to thread slot 3." },
+        { id: "thread4", label: "Thread 4", description: "Jump to thread slot 4." },
+        { id: "thread5", label: "Thread 5", description: "Jump to thread slot 5." },
+        { id: "thread6", label: "Thread 6", description: "Jump to thread slot 6." },
+        { id: "thread7", label: "Thread 7", description: "Jump to thread slot 7." },
+        { id: "thread8", label: "Thread 8", description: "Jump to thread slot 8." },
+        { id: "thread9", label: "Thread 9", description: "Jump to thread slot 9." },
+      ],
+    },
+    {
+      title: "Panels",
+      actions: [
+        { id: "toggleSidebar", label: "Toggle sidebar", description: "Show or hide the sidebar." },
+        { id: "toggleTerminal", label: "Toggle terminal", description: "Show or hide the terminal." },
+        { id: "toggleFileTreePanel", label: "Toggle file tree", description: "Show or hide the file tree." },
+        { id: "openBrowserTab", label: "Open browser tab", description: "Open a browser tab." },
+        { id: "reloadBrowserPage", label: "Reload browser page", description: "Reload the active browser page." },
+        { id: "hardReloadBrowserPage", label: "Hard reload browser page", description: "Hard reload the active browser page." },
+        { id: "toggleBrowserPanel", label: "Toggle browser panel", description: "Show or hide the browser panel." },
+        { id: "toggleDiffPanel", label: "Toggle review panel", description: "Show or hide the review panel." },
+        { id: "openThreadOverlay", label: "Open thread switcher", description: "Open the thread switcher." },
+        { id: "openAvatarOverlay", label: "Open account menu", description: "Open the account menu." },
+      ],
+    },
+    {
+      title: "System",
+      actions: [
+        { id: "toggleTraceRecording", label: "Toggle trace recording", description: "Start or stop trace recording." },
+        { id: "dictation", label: "Dictation", description: "Start dictation." },
+      ],
+    },
+  ];
+
+  return `import{s as __toESM}from"./${chunkAsset}";import{t as __reactFactory}from"./${reactAsset}";import{t as __jsxFactory}from"./${jsxRuntimeAsset}";import{n as __post,rn as DEFAULT_SHORTCUTS}from"./${vscodeApiAsset}";import{i as HotkeyWindowHotkeyRow}from"./${hotkeySettingsAsset}";import{t as Toggle}from"./${toggleAsset}";import{n as SettingsRow}from"./${settingsRowAsset}";import{r as SettingsSection,n as SettingsGroup,t as SettingsPage}from"./${settingsLayoutAsset}";var React=__toESM(__reactFactory(),1),$=__jsxFactory(),KEYS={promptWindow:${JSON.stringify(linuxSettingsKeys.promptWindow)},systemTray:${JSON.stringify(linuxSettingsKeys.systemTray)},warmStart:${JSON.stringify(linuxSettingsKeys.warmStart)}},KEYBIND_OVERRIDES_KEY=${JSON.stringify(linuxKeybindOverridesKey)},KEYBIND_GROUPS=${JSON.stringify(keybindGroups)};function normalizeOverrides(value){if(!value||typeof value!="object"||Array.isArray(value))return{};return Object.fromEntries(Object.entries(value).filter(([key,accelerator])=>typeof key=="string"&&typeof accelerator=="string"&&accelerator.trim().length>0).map(([key,accelerator])=>[key,accelerator.trim()]))}function readLocalOverrides(){try{return normalizeOverrides(JSON.parse(localStorage.getItem(KEYBIND_OVERRIDES_KEY)||"{}"))}catch{return{}}}function writeLocalOverrides(next){try{localStorage.setItem(KEYBIND_OVERRIDES_KEY,JSON.stringify(next)),window.dispatchEvent(new CustomEvent("codex-linux-keybind-overrides-changed",{detail:next}))}catch{}}function useKeybindOverrides(){let[overrides,setOverrides]=React.useState(()=>readLocalOverrides()),[error,setError]=React.useState(null);React.useEffect(()=>{let alive=!0;__post("get-global-state",{params:{key:KEYBIND_OVERRIDES_KEY}}).then(result=>{if(!alive)return;let next=normalizeOverrides(result?.value);Object.keys(next).length>0?(setOverrides(next),writeLocalOverrides(next)):setOverrides(readLocalOverrides());setError(null)}).catch(err=>{alive&&setError(err instanceof Error?err.message:String(err))});return()=>{alive=!1}},[]);let update=React.useCallback((actionId,accelerator)=>{setOverrides(previous=>{let next={...previous},defaultValue=typeof DEFAULT_SHORTCUTS[actionId]=="string"?DEFAULT_SHORTCUTS[actionId]:"",trimmed=String(accelerator??"").trim();trimmed.length===0||trimmed===defaultValue?delete next[actionId]:next[actionId]=trimmed;writeLocalOverrides(next);__post("set-global-state",{params:{key:KEYBIND_OVERRIDES_KEY,value:next}}).then(()=>setError(null)).catch(err=>setError(err instanceof Error?err.message:String(err)));return next})},[]);return{overrides,error,update}}function useLinuxSetting(key,defaultValue){let[value,setValue]=React.useState(defaultValue),[isLoading,setIsLoading]=React.useState(!0),[error,setError]=React.useState(null);React.useEffect(()=>{let alive=!0;setIsLoading(!0);__post("get-global-state",{params:{key}}).then(result=>{alive&&(setValue(result?.value??defaultValue),setError(null))}).catch(err=>{alive&&setError(err instanceof Error?err.message:String(err))}).finally(()=>{alive&&setIsLoading(!1)});return()=>{alive=!1}},[key,defaultValue]);let update=React.useCallback(next=>{let previous=value;setValue(next);setError(null);__post("set-global-state",{params:{key,value:next}}).catch(err=>{setValue(previous);setError(err instanceof Error?err.message:String(err))})},[key,value]);return{value,isLoading,error,update}}function LinuxToggle({settingKey,label,description,defaultValue=!0}){let{value,isLoading,error,update}=useLinuxSetting(settingKey,defaultValue),details=error?$.jsxs("div",{className:"flex flex-col gap-1",children:[$.jsx("span",{children:description}),$.jsx("span",{className:"text-token-error-foreground",children:error})]}):description;return $.jsx(SettingsRow,{label,description:details,control:$.jsx(Toggle,{checked:value,disabled:isLoading,onChange:update,ariaLabel:label})})}function normalizeCapturedKey(key){let map={" ":"Space",ArrowUp:"Up",ArrowDown:"Down",ArrowLeft:"Left",ArrowRight:"Right",Escape:"Esc",",":",",".":".","/":"/","\\\\":"\\\\","[":"[","]":"]",";":";","'":"'","-":"-","=":"=","+":"Plus"};if(map[key])return map[key];if(/^.$/.test(key))return key.toUpperCase();return key}function formatAcceleratorForInput(event){if(!(event.ctrlKey||event.altKey||event.metaKey))return null;if(["Control","Shift","Alt","Meta"].includes(event.key))return null;let parts=[];event.ctrlKey&&parts.push("Ctrl");event.altKey&&parts.push("Alt");event.shiftKey&&parts.push("Shift");event.metaKey&&parts.push("Command");let key=normalizeCapturedKey(event.key);return key?[...parts,key].join("+"):null}function ShortcutInput({value,defaultValue,changed,onChange}){let[draft,setDraft]=React.useState(value);React.useEffect(()=>setDraft(value),[value]);let commit=next=>onChange(String(next??"").trim());return $.jsxs("div",{className:"flex min-w-[260px] items-center justify-end gap-2",children:[$.jsx("input",{className:"h-8 w-[190px] rounded-md border border-token-border-default bg-token-bg-primary px-2 text-sm text-token-text-primary outline-none focus:border-token-border-strong","data-codex-keybind-input":!0,value:draft,placeholder:defaultValue,onChange:event=>{setDraft(event.target.value),onChange(event.target.value)},onBlur:()=>commit(draft),onKeyDown:event=>{if(event.key==="Escape"){setDraft(value);return}if(event.key==="Enter"){event.preventDefault(),commit(draft);return}let captured=formatAcceleratorForInput(event);captured&&(event.preventDefault(),setDraft(captured),onChange(captured))}}),$.jsx("button",{type:"button",className:"h-8 rounded-md border border-token-border-default px-2 text-xs text-token-text-secondary disabled:opacity-40",disabled:!changed,onClick:()=>onChange(""),children:"Reset"})]})}function KeybindRow({action,overrides,update}){let defaultValue=typeof DEFAULT_SHORTCUTS[action.id]=="string"?DEFAULT_SHORTCUTS[action.id]:action.defaultAccelerator??"",hasOverride=Object.prototype.hasOwnProperty.call(overrides,action.id),value=hasOverride?overrides[action.id]:defaultValue,changed=hasOverride&&value!==defaultValue,description=$.jsxs("div",{className:"flex flex-col gap-1",children:[$.jsx("span",{children:action.description}),$.jsxs("span",{className:"text-token-text-tertiary",children:["Default: ",defaultValue||"Unassigned"]})]});return $.jsx(SettingsRow,{label:action.label,description,control:$.jsx(ShortcutInput,{value,defaultValue,changed,onChange:next=>update(action.id,next)})})}function KeybindGroup({group,overrides,update}){return $.jsxs(SettingsSection,{className:"gap-2",children:[$.jsx(SettingsSection.Header,{title:group.title}),$.jsx(SettingsSection.Content,{children:$.jsx(SettingsGroup,{children:group.actions.map(action=>$.jsx(KeybindRow,{action,overrides,update},action.id))})})]},group.title)}function KeybindsSettings(){let{overrides,error,update}=useKeybindOverrides();return $.jsx(SettingsPage,{title:"Keybinds",subtitle:"App shortcuts and Linux desktop behavior.",children:$.jsxs("div",{className:"flex flex-col gap-6",children:[$.jsxs(SettingsSection,{className:"gap-2",children:[$.jsx(SettingsSection.Header,{title:"App shortcuts"}),error?$.jsx("div",{className:"px-1 text-sm text-token-error-foreground",children:error}):null]}),...KEYBIND_GROUPS.map(group=>$.jsx(KeybindGroup,{group,overrides,update},group.title)),$.jsxs(SettingsSection,{className:"gap-2",children:[$.jsx(SettingsSection.Header,{title:"Global shortcuts"}),$.jsx(SettingsSection.Content,{children:$.jsxs(SettingsGroup,{children:[$.jsx(HotkeyWindowHotkeyRow,{}),$.jsx(LinuxToggle,{settingKey:KEYS.promptWindow,label:"Compact prompt window",description:"Allow --prompt-chat and --hotkey-window to open the compact prompt window and keep it prewarmed."})]})})]}),$.jsxs(SettingsSection,{className:"gap-2",children:[$.jsx(SettingsSection.Header,{title:"Linux desktop"}),$.jsx(SettingsSection.Content,{children:$.jsxs(SettingsGroup,{children:[$.jsx(LinuxToggle,{settingKey:KEYS.systemTray,label:"System tray",description:"Show the Codex system tray icon and keep the app available from the tray."}),$.jsx(LinuxToggle,{settingKey:KEYS.warmStart,label:"Warm start",description:"Use the running app for launch actions instead of starting a fresh Electron instance."})]})})]})]})})}export{KeybindsSettings,KeybindsSettings as default};\n//# sourceMappingURL=${keybindsSettingsAsset}.map\n`;
+}
+
+function resolveKeybindsSettingsAsset(extractedDir) {
+  const webviewAssetsDir = path.join(extractedDir, "webview", "assets");
+  if (!fs.existsSync(webviewAssetsDir)) {
+    throw new Error(`Required Keybinds settings patch failed: missing webview assets directory ${webviewAssetsDir}`);
+  }
+
+  const reactAsset = findRequiredWebviewAsset(webviewAssetsDir, /^react-.*\.js$/, "react.transitional.element", "React asset");
+  const chunkAsset = findImportedAsset(webviewAssetsDir, reactAsset, "React shared chunk asset");
+  const jsxRuntimeAsset = findRequiredWebviewAsset(webviewAssetsDir, /^jsx-runtime-.*\.js$/, "react.transitional.element", "JSX runtime asset");
+  const vscodeApiAsset = findRequiredWebviewAsset(webviewAssetsDir, /^vscode-api-.*\.js$/, "vscode://codex", "VS Code API asset");
+  const hotkeySettingsAsset = findRequiredWebviewAsset(
+    webviewAssetsDir,
+    /^general-settings-.*\.js$/,
+    "hotkey-window-hotkey-state",
+    "hotkey settings asset",
+  );
+  const toggleAsset = findRequiredWebviewAsset(webviewAssetsDir, /^toggle-.*\.js$/, null, "toggle asset");
+  const settingsRowAsset = findRequiredWebviewAsset(webviewAssetsDir, /^settings-row-.*\.js$/, null, "settings row asset");
+  const settingsLayoutAsset = findRequiredWebviewAsset(
+    webviewAssetsDir,
+    /^settings-content-layout-.*\.js$/,
+    null,
+    "settings content layout asset",
+  );
+  const filePath = path.join(webviewAssetsDir, keybindsSettingsAsset);
+
+  return {
+    filePath,
+    source: buildKeybindsSettingsSource({
+      chunkAsset,
+      reactAsset,
+      jsxRuntimeAsset,
+      vscodeApiAsset,
+      hotkeySettingsAsset,
+      toggleAsset,
+      settingsRowAsset,
+      settingsLayoutAsset,
+    }),
+  };
+}
+
+function collectRequiredAssetPatches(extractedDir, filenamePattern, patchFn, description) {
+  const webviewAssetsDir = path.join(extractedDir, "webview", "assets");
+  if (!fs.existsSync(webviewAssetsDir)) {
+    throw new Error(`Required Keybinds settings patch failed: missing webview assets directory ${webviewAssetsDir}`);
+  }
+
+  const candidates = fs
+    .readdirSync(webviewAssetsDir)
+    .filter((name) => filenamePattern.test(name))
+    .sort();
+  if (candidates.length === 0) {
+    throw new Error(`Required Keybinds settings patch failed: could not find ${description}`);
+  }
+
+  return candidates.map((candidate) => {
+    const filePath = path.join(webviewAssetsDir, candidate);
+    const currentSource = fs.readFileSync(filePath, "utf8");
+    return {
+      filePath,
+      currentSource,
+      patchedSource: patchFn(currentSource),
+    };
+  });
+}
+
+function patchKeybindsSettingsAssets(extractedDir) {
+  try {
+    const keybindsAsset = resolveKeybindsSettingsAsset(extractedDir);
+    const keybindsAssetExists = fs.existsSync(keybindsAsset.filePath);
+    const previousKeybindsSource = keybindsAssetExists
+      ? fs.readFileSync(keybindsAsset.filePath, "utf8")
+      : null;
+    const patches = [
+      ...collectRequiredAssetPatches(
+        extractedDir,
+        /^settings-sections-.*\.js$/,
+        applyKeybindsSettingsSectionsPatch,
+        "settings sections bundle",
+      ),
+      ...collectRequiredAssetPatches(
+        extractedDir,
+        /^settings-shared-.*\.js$/,
+        applyKeybindsSettingsSharedPatch,
+        "settings shared bundle",
+      ),
+      ...collectRequiredAssetPatches(
+        extractedDir,
+        /^index-.*\.js$/,
+        applyKeybindsSettingsIndexPatch,
+        "webview index bundle",
+      ),
+    ];
+
+    fs.writeFileSync(keybindsAsset.filePath, keybindsAsset.source, "utf8");
+    let changed = previousKeybindsSource !== keybindsAsset.source ? 1 : 0;
+    for (const patch of patches) {
+      if (patch.patchedSource !== patch.currentSource) {
+        fs.writeFileSync(patch.filePath, patch.patchedSource, "utf8");
+        changed += 1;
+      }
+    }
+    return { matched: true, changed };
+  } catch (error) {
+    const message = error instanceof Error ? error.message : String(error);
+    console.warn(`WARN: Keybinds settings patch skipped: ${message}`);
+    return { matched: false, changed: 0, reason: message };
+  }
+}
+
+function applyKeybindsSettingsSectionsPatch(currentSource) {
+  let patchedSource = currentSource;
+
+  if (patchedSource.includes("slug:`keybinds`")) {
+    return patchedSource;
+  }
+
+  const sectionsNeedle = "var e=`general-settings`,t=`mcp-settings`,n=[{slug:e},";
+  const sectionsPatch = "var e=`general-settings`,t=`mcp-settings`,n=[{slug:e},{slug:`keybinds`},";
+  if (patchedSource.includes(sectionsNeedle)) {
+    return patchedSource.replace(sectionsNeedle, sectionsPatch);
+  }
+
+  const currentNeedle = "n=[{slug:e},{slug:`appearance`}";
+  if (patchedSource.includes(currentNeedle)) {
+    return patchedSource.replace(currentNeedle, "n=[{slug:e},{slug:`keybinds`},{slug:`appearance`}");
+  }
+
+  const literalNeedle = "n=[{slug:`general-settings`},{slug:`appearance`}";
+  if (patchedSource.includes(literalNeedle)) {
+    return patchedSource.replace(literalNeedle, "n=[{slug:`general-settings`},{slug:`keybinds`},{slug:`appearance`}");
+  }
+
+  throw new Error("Required Keybinds settings patch failed: could not add keybinds settings section");
+}
+
+function applyKeybindsSettingsSharedPatch(currentSource) {
+  let patchedSource = currentSource;
+
+  if (!patchedSource.includes("settings.nav.keybinds")) {
+    const navNeedle =
+      '"general-settings":{id:`settings.nav.general-settings`,defaultMessage:`General`,description:`Title for general settings section`},';
+    const navPatch =
+      '"general-settings":{id:`settings.nav.general-settings`,defaultMessage:`General`,description:`Title for general settings section`},keybinds:{id:`settings.nav.keybinds`,defaultMessage:`Keybinds`,description:`Title for keybinds settings section`},';
+    if (!patchedSource.includes(navNeedle)) {
+      throw new Error("Required Keybinds settings patch failed: could not add keybinds nav label");
+    }
+    patchedSource = patchedSource.replace(navNeedle, navPatch);
+  }
+
+  if (!patchedSource.includes("settings.section.keybinds")) {
+    const sectionNeedle =
+      "case`general-settings`:{let e;return t[2]===Symbol.for(`react.memo_cache_sentinel`)?(e=(0,d.jsx)(n,{id:`settings.section.general-settings`,defaultMessage:`General`,description:`Title for general settings section`}),t[2]=e):e=t[2],e}";
+    const sectionPatch =
+      "case`general-settings`:{let e;return t[2]===Symbol.for(`react.memo_cache_sentinel`)?(e=(0,d.jsx)(n,{id:`settings.section.general-settings`,defaultMessage:`General`,description:`Title for general settings section`}),t[2]=e):e=t[2],e}case`keybinds`:{return (0,d.jsx)(n,{id:`settings.section.keybinds`,defaultMessage:`Keybinds`,description:`Title for keybinds settings section`})}";
+    if (!patchedSource.includes(sectionNeedle)) {
+      throw new Error("Required Keybinds settings patch failed: could not add keybinds section title");
+    }
+    patchedSource = patchedSource.replace(sectionNeedle, sectionPatch);
+  }
+
+  return patchedSource;
+}
+
+function applyLinuxKeybindOverridesRuntimePatch(currentSource) {
+  const runtimePatch = `;function codexLinuxKeybindOverridesRuntime(){try{if(typeof window=="undefined")return;let storageKey=${JSON.stringify(linuxKeybindOverridesKey)},defaultMap=typeof Ct=="object"&&Ct?Ct:{},overrides={};function loadOverrides(){try{let value=JSON.parse(localStorage.getItem(storageKey)||"{}");overrides=value&&typeof value=="object"&&!Array.isArray(value)?value:{}}catch{overrides={}}}function isShortcutCaptureTarget(event){let target=event.target;return target instanceof Element&&target.closest("[data-codex-keybind-input]")!=null}function normalizeKeyName(key){let map={Space:" ",Esc:"Escape",Up:"ArrowUp",Down:"ArrowDown",Left:"ArrowLeft",Right:"ArrowRight",Plus:"+",Comma:",",Period:".",Slash:"/"};return map[key]??(/^.$/.test(key)?key.toUpperCase():key)}function parseAccelerator(accelerator){if(typeof accelerator!="string"||accelerator.trim().length===0)return null;let isMac=/Mac/.test(navigator.platform||""),parts=accelerator.split("+").map(part=>part.trim()).filter(Boolean),parsed={ctrl:false,alt:false,shift:false,meta:false,key:null};for(let part of parts){switch(part){case"CmdOrCtrl":isMac?parsed.meta=true:parsed.ctrl=true;break;case"Command":case"Cmd":case"Meta":case"Super":case"Win":parsed.meta=true;break;case"Control":case"Ctrl":parsed.ctrl=true;break;case"Alt":case"Option":parsed.alt=true;break;case"Shift":parsed.shift=true;break;default:parsed.key=normalizeKeyName(part);break}}return parsed.key?parsed:null}function matches(event,parsed){return event.ctrlKey===parsed.ctrl&&event.altKey===parsed.alt&&event.shiftKey===parsed.shift&&event.metaKey===parsed.meta&&normalizeKeyName(event.key)===parsed.key}function dispatchHost(message){if(typeof E=="object"&&E&&typeof E.dispatchHostMessage=="function"){E.dispatchHostMessage(message);return true}return false}function dispatchElectron(type,params={}){if(typeof E=="object"&&E&&typeof E.dispatchMessage=="function"){E.dispatchMessage(type,params);return true}return false}let hostActionTypes={newThread:"new-chat",quickChat:"new-quick-chat",newThreadAlt:"new-chat",toggleSidebar:"toggle-sidebar",toggleTerminal:"toggle-terminal",toggleBrowserPanel:"toggle-browser-panel",toggleDiffPanel:"toggle-diff-panel",findInThread:"find-in-thread",navigateBack:"navigate-back",navigateForward:"navigate-forward",previousThread:"previous-thread",nextThread:"next-thread",copyConversationPath:"copy-conversation-path",toggleThreadPin:"toggle-thread-pin",renameThread:"rename-thread",archiveThread:"archive-thread",copyWorkingDirectory:"copy-working-directory",copySessionId:"copy-session-id",copyDeeplink:"copy-deeplink",toggleFileTreePanel:"toggle-file-tree-panel"};function runAction(id){if(/^thread[1-9]$/.test(id))return dispatchHost({type:"go-to-thread-index",index:Number(id.slice(6))-1});switch(id){case"openCommandMenu":case"openCommandMenuAlt":return dispatchHost({type:"command-menu",query:""});case"searchChats":return dispatchHost({type:"chat-search-command-menu"});case"searchFiles":return dispatchHost({type:"file-search-command-menu"});case"openFolder":return dispatchElectron("electron-create-new-workspace-root-option",{});case"settings":return dispatchElectron("show-settings",{section:"general-settings"});case"openBrowserTab":return dispatchHost({type:"browser-sidebar-command",command:{type:"new-tab"}});case"reloadBrowserPage":return dispatchHost({type:"browser-sidebar-command",command:{type:"reload"}});case"hardReloadBrowserPage":return dispatchHost({type:"browser-sidebar-command",command:{type:"hard-reload"}});case"dictation":return dispatchElectron("global-dictation-start",{});default:return hostActionTypes[id]?dispatchHost({type:hostActionTypes[id]}):false}}loadOverrides();window.addEventListener("storage",event=>{event.key===storageKey&&loadOverrides()});window.addEventListener("codex-linux-keybind-overrides-changed",loadOverrides);window.addEventListener("keydown",event=>{if(event.defaultPrevented||event.repeat||isShortcutCaptureTarget(event))return;for(let[id,accelerator]of Object.entries(overrides)){if(typeof accelerator!="string"||accelerator.trim().length===0||accelerator.trim()===(defaultMap[id]||""))continue;let parsed=parseAccelerator(accelerator);if(parsed&&matches(event,parsed)&&runAction(id)){event.preventDefault();event.stopPropagation();break}}},true)}catch{}}codexLinuxKeybindOverridesRuntime();`;
+
+  const runtimeMarker = ";function codexLinuxKeybindOverridesRuntime()";
+  const existingRuntimeIndex = currentSource.indexOf(runtimeMarker);
+  if (existingRuntimeIndex !== -1) {
+    return `${currentSource.slice(0, existingRuntimeIndex).trimEnd()}\n${runtimePatch}`;
+  }
+
+  return `${currentSource}\n${runtimePatch}`;
+}
+
+function applyKeybindsSettingsIndexPatch(currentSource) {
+  let patchedSource = currentSource;
+
+  if (!patchedSource.includes(`${keybindsSettingsAsset}`)) {
+    const routePattern = /var ([A-Za-z_$][\w$]*)=\{"general-settings":(?=\(0,[A-Za-z_$][\w$]*\.lazy\))/;
+    if (!routePattern.test(patchedSource)) {
+      throw new Error("Required Keybinds settings patch failed: could not add keybinds route");
+    }
+    patchedSource = patchedSource.replace(
+      routePattern,
+      `var $1={keybinds:(0,Z.lazy)(()=>s(()=>import(\`./${keybindsSettingsAsset}\`),[],import.meta.url)),"general-settings":`,
+    );
+  }
+
+  if (!/[,{]keybinds:[A-Za-z_$][\w$]*,"general-settings":/.test(patchedSource)) {
+    const iconPattern = /([A-Za-z_$][\w$]*=\{)"general-settings":([A-Za-z_$][\w$]*),/;
+    if (!iconPattern.test(patchedSource)) {
+      throw new Error("Required Keybinds settings patch failed: could not add keybinds icon");
+    }
+    patchedSource = patchedSource.replace(
+      iconPattern,
+      (_match, prefix, icon) => `${prefix}keybinds:${icon},"general-settings":${icon},`,
+    );
+  }
+
+  if (!/=\[`general-settings`,`keybinds`/.test(patchedSource)) {
+    const orderPattern = /([A-Za-z_$][\w$]*=\[`general-settings`,)`appearance`/;
+    if (!orderPattern.test(patchedSource)) {
+      throw new Error("Required Keybinds settings patch failed: could not add keybinds nav order");
+    }
+    patchedSource = patchedSource.replace(orderPattern, "$1`keybinds`,`appearance`");
+  }
+
+  if (!patchedSource.includes("slugs:[`general-settings`,`keybinds`")) {
+    const groupNeedle = "slugs:[`general-settings`,`appearance`,`connections`,`git-settings`,`usage`]";
+    const groupPatch = "slugs:[`general-settings`,`keybinds`,`appearance`,`connections`,`git-settings`,`usage`]";
+    if (!patchedSource.includes(groupNeedle)) {
+      throw new Error("Required Keybinds settings patch failed: could not add keybinds nav group");
+    }
+    patchedSource = patchedSource.replace(groupNeedle, groupPatch);
+  }
+
+  if (!patchedSource.includes("case`keybinds`:return l===`electron`")) {
+    const visibilityNeedle =
+      "case`appearance`:case`git-settings`:case`worktrees`:case`local-environments`:case`data-controls`:case`environments`:return l===`electron`;";
+    const visibilityPatch =
+      "case`keybinds`:return l===`electron`;case`appearance`:case`git-settings`:case`worktrees`:case`local-environments`:case`data-controls`:case`environments`:return l===`electron`;";
+    if (!patchedSource.includes(visibilityNeedle)) {
+      throw new Error("Required Keybinds settings patch failed: could not add keybinds visibility");
+    }
+    patchedSource = patchedSource.replace(visibilityNeedle, visibilityPatch);
+  }
+
+  if (!patchedSource.includes("case`keybinds`:k=!1;break bb0;")) {
+    const redirectNeedle =
+      "case`appearance`:case`general-settings`:case`agent`:case`git-settings`:case`account`:case`data-controls`:case`personalization`:k=!1;break bb0;";
+    const redirectPatch =
+      "case`keybinds`:k=!1;break bb0;case`appearance`:case`general-settings`:case`agent`:case`git-settings`:case`account`:case`data-controls`:case`personalization`:k=!1;break bb0;";
+    if (patchedSource.includes(redirectNeedle)) {
+      patchedSource = patchedSource.replace(redirectNeedle, redirectPatch);
+    }
+  }
+
+  return applyLinuxKeybindOverridesRuntimePatch(patchedSource);
+}
+
+module.exports = {
+  applyKeybindsSettingsIndexPatch,
+  applyKeybindsSettingsSectionsPatch,
+  applyKeybindsSettingsSharedPatch,
+  applyLinuxKeybindOverridesRuntimePatch,
+  keybindsSettingsAsset,
+  linuxKeybindOverridesKey,
+  patchKeybindsSettingsAssets,
+  resolveKeybindsSettingsAsset,
+};
