+++ scripts/patches/webview-assets.js
@@ -1,0 +1,139 @@
+"use strict";
+
+const fs = require("node:fs");
+const path = require("node:path");
+
+// Webview asset patches target hashed browser chunks copied out of app.asar.
+// They stay fail-soft because upstream chunk names and minified symbols drift.
+function applyLinuxOpaqueWindowsDefaultPatch(currentSource) {
+  let patchedSource = currentSource;
+
+  const mergeNeedle = "opaqueWindows:e?.opaqueWindows??n.opaqueWindows,semanticColors:";
+  const mergePatch =
+    "opaqueWindows:e?.opaqueWindows??(typeof navigator<`u`&&((navigator.userAgentData?.platform??navigator.platform??navigator.userAgent).toLowerCase().includes(`linux`))?!0:n.opaqueWindows),semanticColors:";
+
+  if (patchedSource.includes("opaqueWindows:e?.opaqueWindows??(typeof navigator<`u`&&")) {
+    // Already patched.
+  } else if (patchedSource.includes(mergeNeedle)) {
+    patchedSource = patchedSource.replace(mergeNeedle, mergePatch);
+  } else if (patchedSource.includes("opaqueWindows") && patchedSource.includes("semanticColors")) {
+    console.warn(
+      "WARN: Could not find Linux opaque window default insertion point — skipping settings default patch",
+    );
+  }
+
+  const settingsNeedle =
+    "let d=ot(r,e),f=at(e),p={codeThemeId:tt(a,e).id,theme:d},";
+  const settingsPatch =
+    "let d=ot(r,e);navigator.userAgent.includes(`Linux`)&&r?.opaqueWindows==null&&(d={...d,opaqueWindows:!0});let f=at(e),p={codeThemeId:tt(a,e).id,theme:d},";
+  if (patchedSource.includes("navigator.userAgent.includes(`Linux`)&&r?.opaqueWindows==null")) {
+    // Already patched.
+  } else if (patchedSource.includes(settingsNeedle)) {
+    patchedSource = patchedSource.replace(settingsNeedle, settingsPatch);
+  }
+
+  const currentSettingsNeedle = "setThemePatch:b,theme:x}=ne(t),S=$t(i,t),";
+  const currentSettingsPatch =
+    "setThemePatch:b,theme:x}=ne(t);navigator.userAgent.includes(`Linux`)&&x?.opaqueWindows==null&&(x={...x,opaqueWindows:!0});let S=$t(i,t),";
+  if (patchedSource.includes("navigator.userAgent.includes(`Linux`)&&x?.opaqueWindows==null")) {
+    // Already patched.
+  } else if (patchedSource.includes(currentSettingsNeedle)) {
+    patchedSource = patchedSource.replace(currentSettingsNeedle, currentSettingsPatch);
+  }
+
+  const runtimeNeedle =
+    "let T=o===`light`?C:w,E;if(T.opaqueWindows&&!XZ()){";
+  const runtimePatch =
+    "let T=o===`light`?C:w,E;document.documentElement.dataset.codexOs===`linux`&&((o===`light`?l:f)?.opaqueWindows==null&&(T={...T,opaqueWindows:!0}));if(T.opaqueWindows&&!XZ()){";
+  if (patchedSource.includes("document.documentElement.dataset.codexOs===`linux`&&((o===`light`?l:f)?.opaqueWindows==null")) {
+    // Already patched.
+  } else if (patchedSource.includes(runtimeNeedle)) {
+    patchedSource = patchedSource.replace(runtimeNeedle, runtimePatch);
+  }
+
+  const currentRuntimeNeedle = "let T=s===`light`?S:w,E;";
+  const currentRuntimePatch =
+    "let T=s===`light`?S:w,E;document.documentElement.dataset.codexOs===`linux`&&((s===`light`?u:p)?.opaqueWindows==null&&(T={...T,opaqueWindows:!0}));";
+  if (patchedSource.includes("document.documentElement.dataset.codexOs===`linux`&&((s===`light`?u:p)?.opaqueWindows==null")) {
+    // Already patched.
+  } else if (patchedSource.includes(currentRuntimeNeedle)) {
+    patchedSource = patchedSource.replace(currentRuntimeNeedle, currentRuntimePatch);
+  }
+
+  return patchedSource;
+}
+
+function applyLinuxAppSunsetPatch(currentSource) {
+  const statsigKey = "2929582856";
+  const disabledGatePattern = /if\(!1&&([A-Za-z_$][\w$]*)\(`2929582856`\)\)\{/u;
+  const gatePattern = /if\(([A-Za-z_$][\w$]*)\(`2929582856`\)\)\{/u;
+
+  if (disabledGatePattern.test(currentSource)) {
+    return currentSource;
+  }
+
+  if (gatePattern.test(currentSource)) {
+    return currentSource.replace(gatePattern, "if(!1&&$1(`2929582856`)){");
+  }
+
+  if (currentSource.includes(statsigKey)) {
+    console.warn("WARN: Could not find app sunset gate needle — skipping Linux app sunset patch");
+  }
+
+  return currentSource;
+}
+
+function applyBrowserAnnotationScreenshotPatch(currentSource) {
+  let patchedSource = currentSource;
+
+  const liveElementScreenshotNeedle =
+    "if(M&&j?.anchor.kind===`element`){let e=qu(j,y.current)??null,t=e==null?null:rd(e);he=t?.rect??md(j.anchor),_e=t?.borderRadius}";
+  const storedAnchorScreenshotPatch =
+    "if(M&&j?.anchor.kind===`element`){he=md(j.anchor),_e=void 0}";
+  if (patchedSource.includes(storedAnchorScreenshotPatch)) {
+    // Already patched.
+  } else if (patchedSource.includes(liveElementScreenshotNeedle)) {
+    patchedSource = patchedSource.replace(liveElementScreenshotNeedle, storedAnchorScreenshotPatch);
+  } else {
+    console.warn("WARN: Could not find browser annotation screenshot element highlight — skipping screenshot anchor patch");
+  }
+
+  const allMarkersInScreenshotNeedle =
+    "de=u?.target.mode===`create`?ce.find(e=>Sd(e.anchor,u.anchor.value))??null:null,fe=!M&&de!=null?ce.filter(e=>e.id!==de.id):ce,";
+  const selectedMarkerInScreenshotPatch =
+    "de=u?.target.mode===`create`?ce.find(e=>Sd(e.anchor,u.anchor.value))??null:null,fe=M?ue:!M&&de!=null?ce.filter(e=>e.id!==de.id):ce,";
+  if (patchedSource.includes(selectedMarkerInScreenshotPatch)) {
+    // Already patched.
+  } else if (patchedSource.includes(allMarkersInScreenshotNeedle)) {
+    patchedSource = patchedSource.replace(allMarkersInScreenshotNeedle, selectedMarkerInScreenshotPatch);
+  } else {
+    console.warn("WARN: Could not find browser annotation screenshot markers — skipping screenshot marker patch");
+  }
+
+  return patchedSource;
+}
+
+function patchCommentPreloadBundle(extractedDir) {
+  const commentPreloadBundle = path.join(extractedDir, ".vite", "build", "comment-preload.js");
+  if (!fs.existsSync(commentPreloadBundle)) {
+    console.warn(
+      `WARN: Could not find comment preload bundle in ${path.dirname(commentPreloadBundle)} — skipping annotation screenshot patch`,
+    );
+    return { matched: false, changed: false };
+  }
+
+  const source = fs.readFileSync(commentPreloadBundle, "utf8");
+  const patchedSource = applyBrowserAnnotationScreenshotPatch(source);
+  if (patchedSource !== source) {
+    fs.writeFileSync(commentPreloadBundle, patchedSource, "utf8");
+    return { matched: true, changed: true };
+  }
+  return { matched: true, changed: false };
+}
+
+module.exports = {
+  applyBrowserAnnotationScreenshotPatch,
+  applyLinuxAppSunsetPatch,
+  applyLinuxOpaqueWindowsDefaultPatch,
+  patchCommentPreloadBundle,
+};
